Nest classes in partial SpriteText class
